applications/task-2-rag-system/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.0.3
langchain==0.1.0
openai==1.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
pinecone-client==3.0.0
pypdf==3.17.0
aiofiles==23.2.1
httpx==0.25.0
websockets==12.0
